chore: 유저 프로퍼티 타입 변경
