+++ src/components/auth/types/userTypes.ts
@@ -1,0 +1,22 @@
+export interface RegisterType {
+  name: string
+  readname: string
+  postNumber:string;
+  address: string
+  phone: string
+  email: string
+  birthday: string
+  password: string
+  recommandCode: string;
+}
+
+export interface UserDataType {
+  name: string;
+  readname:string;
+  postNumber:string;
+  address: string;
+  phone: string;
+  email:string;
+  birthday:string;
+  token: string;
+}
